% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{json2env-package}
\alias{json2env}
\alias{json2env-package}
\title{json2env: Parse JSON Formatted Arguments}
\description{
Parses an argument in JSON format into arguments for the enclosing function.
}
\details{
\code{\link{json2env}} is the only function in the package json2env.
If this function proves useful, it could be added to another package
such as \code{\link{jsonlite}} or in a language parsing utility package.
}
\keyword{internal}
